[2018.04.21] update

+++ lib/gcforest/utils/win_utils.py
@@ -41,7 +41,7 @@ def get_windows(X, win_x, win_y, stride_x=1, stride_y=1, pad_x=0, pad_y=0):
     nc = win_y * win_x * c
     nh = (h - win_y) / stride_y + 1
     nw = (w - win_x) / stride_x + 1
-    X_win = np.empty(( nc, n * nh * nw), dtype=np.float32)
+    X_win = np.empty(( nc, n * nh * nw), dtype=X.dtype)
     LOGGER.info("get_windows_start: X.shape={}, X_win.shape={}, nw={}, nh={}, c={}, win_x={}, win_y={}, stride_x={}, stride_y={}".format(
                 X.shape, X_win.shape, nw, nh, c, win_x, win_y, stride_x, stride_y))
     Parallel(n_jobs=-1, backend="threading", verbose=0)(
